requirements.txt
pywinrm==0.2.2
paramiko
Flask
boto3